feat!: refactor code again
This redoes a lot of the config flow and sensor details and adds proper sensors for the various aspects of the library that one might want to track.
It also better handles errors and implements automated releases.
I ended up switching to a `requirements.txt` for pip just because it's easier to manage than poetry

There is the start of support for translations and testing as well.